xml-docs for telegram and twitter namespaces

+++ Osnova.Net/Twitter/OsnovaTweet.cs
@@ -9,9 +9,10 @@ using Osnova.Net.JsonConverters;
 namespace Osnova.Net.Twitter
 {
     /// <summary>
-    /// OsnovaShit
+    /// Osnova implementation of <see cref="Tweet"/> v 1.1;
     /// Used for GetTweets method only
-    /// Refers to Tweet spec
+    /// <para/>
+    /// <remarks>Refers to Tweet spec</remarks>
     /// </summary>
     public class OsnovaTweet : Tweet
     {
@@ -19,13 +20,22 @@ namespace Osnova.Net.Twitter
 
         #region Wrong types
 
+        /// <summary>
+        /// Tweet's author
+        /// </summary>
         [JsonPropertyName("user")]
         public new OsnovaTwitterUser User { get; set; }
 
+        /// <summary>
+        /// Tweet's ID
+        /// </summary>
         [JsonNumberHandling(JsonNumberHandling.AllowReadingFromString | JsonNumberHandling.WriteAsString)]
         [JsonPropertyName("id")]
         public new long Id { get; set; }
 
+        /// <summary>
+        /// Date when this tweet was created
+        /// </summary>
         [JsonConverter(typeof(LongDateTimeOffsetJsonConverter))]
         [JsonPropertyName("created_at")]
         public new DateTimeOffset CreatedAt { get; set; }
@@ -34,9 +44,15 @@ namespace Osnova.Net.Twitter
 
         #region Osnova only stuff
 
+        /// <summary>
+        /// Has media?
+        /// </summary>
         [JsonPropertyName("has_media")]
         public bool HasMedia { get; set; }
 
+        /// <summary>
+        /// Collection of <see cref="OsnovaTweetMedia"/> elements, if any
+        /// </summary>
         [JsonPropertyName("media")]
         public IEnumerable<OsnovaTweetMedia> Media { get; set; }
 
@@ -53,8 +69,17 @@ namespace Osnova.Net.Twitter
 
         #region GetTweets
 
+        /// <summary>
+        /// Gets tweets URL
+        /// </summary>
+        /// <param name="websiteKind">Kind of website</param>
+        /// <param name="tweetSorting">Sorting of tweets</param>
+        /// <param name="count">Count of tweets</param>
+        /// <param name="offset">Tweet offset</param>
+        /// <param name="apiVersion">Target version of API</param>
+        /// <returns>Ready URL, e.g.: https://api.dtf.ru/v1.9/tweets/fresh?count=10</returns>
         public static Uri GetTweetsUri(WebsiteKind websiteKind, TweetSorting tweetSorting = TweetSorting.Fresh,
-                                       long count = -1, long offset = -1, double apiVersion = Core.ApiVersion)
+                                       int count = -1, int offset = -1, double apiVersion = Core.ApiVersion)
         {
             var baseUri = Core.GetBaseUri(websiteKind, apiVersion);
 
@@ -68,14 +93,38 @@ namespace Osnova.Net.Twitter
             return builder.Uri;
         }
 
+        /// <summary>
+        /// Gets tweets
+        /// <para/>
+        /// <remarks>Original name: getTweets</remarks>
+        /// </summary>
+        /// <param name="client">Client to send requests</param>
+        /// <param name="websiteKind">Kind of website</param>
+        /// <param name="tweetSorting">Sorting of tweets</param>
+        /// <param name="count">Count of tweets</param>
+        /// <param name="offset">Tweet offset</param>
+        /// <param name="apiVersion">Target version of API</param>
+        /// <returns>Requested tweets</returns>
         public static ValueTask<HttpResponseMessage> GetTweetsResponseAsync(HttpClient client, WebsiteKind websiteKind,
-                TweetSorting tweetSorting = TweetSorting.Fresh, long count = -1, long offset = -1, double apiVersion = Core.ApiVersion)
+                TweetSorting tweetSorting = TweetSorting.Fresh, int count = -1, int offset = -1, double apiVersion = Core.ApiVersion)
         {
             return Core.GetResponseFromApiAsync(client, GetTweetsUri(websiteKind, tweetSorting, count, offset, apiVersion));
         }
 
+        /// <summary>
+        /// Gets tweets
+        /// <para/>
+        /// <remarks>Original name: getTweets</remarks>
+        /// </summary>
+        /// <param name="client">Client to send requests</param>
+        /// <param name="websiteKind">Kind of website</param>
+        /// <param name="tweetSorting">Sorting of tweets</param>
+        /// <param name="count">Count of tweets</param>
+        /// <param name="offset">Tweet offset</param>
+        /// <param name="apiVersion">Target version of API</param>
+        /// <returns>Requested tweets</returns>
         public static async ValueTask<IEnumerable<OsnovaTweet>> GetTweetsAsync(HttpClient client, WebsiteKind websiteKind,
-               TweetSorting tweetSorting = TweetSorting.Fresh, long count = -1, long offset = -1, double apiVersion = Core.ApiVersion)
+               TweetSorting tweetSorting = TweetSorting.Fresh, int count = -1, int offset = -1, double apiVersion = Core.ApiVersion)
         {
             var response = await GetTweetsResponseAsync(client, websiteKind, tweetSorting, count, offset, apiVersion).ConfigureAwait(false);
